--- javascript/code-challenges/stacksAndQueues/README.md
@@ -1,24 +1,0 @@
-# stacks and queues
-
-<!-- Short summary or background information -->
-
-## Challenge
-
-create the base data structure for stacks and queues
-
-
-## Approach & Efficiency
-
-- stack.peak() O()
-- stack.push() O(n)
-- stack.pop() O(n)
-- stack.isEmpty() O(1)
-- queue.enqueue(item) O(1)
-- queue.dequeue() O(n)
-- queue.peak() O(1)
-- queue.isEmpty() O(1)
-
-
-
-## Solution
-<!-- Embedded whiteboard image -->
