--- javascript/code-challenges/stacksAndQueues/node.js
@@ -1,10 +1,0 @@
-'use strict';
-
-class Node {
-  constructor(value){
-    this.value = value;
-    this.next = null;
-  }
-}
-
-module.exports = Node;
