--- javascript/code-challenges/stacksAndQueues/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "stacksAndQueues",
-  "version": "1.0.0",
-  "description": "",
-  "main": "stacks-and-queues.js",
-  "scripts": {
-    "test": "jest --watchAll --coverage"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
