+++ javascript/trees/README.md
@@ -1,0 +1,17 @@
+# Trees
+
+<!-- Short summary or background information -->
+
+## Challenge
+
+Trees data structure
+
+## Approach & Efficiency
+
+followed along with a demo on trees provied by lena,
+
+need to clarify what is going on.
+
+## Solution
+<!-- Embedded whiteboard image -->
+[whiteboard]('./assets/code-challenge16.png');
