+++ javascript/trees/__test__/findMax.test.js
@@ -1,0 +1,3 @@
+// it should find max value any where in a tree
+
+// it should return null on empty tree
