+++ javascript/trees/node.js
@@ -1,0 +1,11 @@
+'use strict';
+
+class Node {
+  constructor(value=0 ,left =null ,right=null ){
+    this.value = value;
+    this.left = left;
+    this.right = right;
+  }
+}
+
+module.exports = Node;
